dmr constant

+++ src/cuda/micro_mp_hard/dmr_nonconstant_kernels.h
@@ -79,8 +79,9 @@ __global__ void MicroBenchmarkKernel_ADDNONCONSTANT(real_t* input,
 
 		if ((count % num_op) == 0) {
 			check_bit_error(acc_half_t, acc_real_t);
+
 			threshold = acc_real_t - real_t(acc_half_t);
-			acc_half_t = acc_real_t;
+			acc_half_t = half_t(acc_real_t);
 		}
 	}
 	output_real_t[thread_id] = acc_real_t;
